requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
contextlib2==21.6.0
Django==5.0.2
django-extensions==3.2.3
django-ninja==1.1.0
django-ninja-extra==0.20.4
injector==0.21.0
pydantic==2.6.3
pydantic_core==2.16.3
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
